A 'get rid of warnings' change.
Suggested by: Greg Stein <gstein@lyra.org>
Review by: Greg Stein <gstein@lyra.org>

* subversion/include/svn_delta.h

  (svn_txdelta_op_t): Change type of 'offset'
    and 'length' from apr_off_t to apr_size_t.


* subversion/libsvn_fs/strings-table.c

  (string_read): Cast 'offset' to the type used
    by Berkeley DB.


* subversion/libsvn_fs/reps-strings.c

  (window_handler): Change type of 't' from
    int to apr_size_t.


* subversion/libsvn_wc/props.c

  (svn_wc__merge_prop_diffs): Change type of
    'len' from apr_off_t to apr_size_t.  Also
    add a typecast.


* subversion/libsvn_wc/log.c

  (svn_wc_cleanup): Use Gregs Real Man solution
    to set 'is_this_dir'.


* subversion/clients/cmdline/util.c

  (svn_cl__edit_externally): Add typecasts.


* subversion/libsvn_delta/svndiff.c

  (encode_int): Add typecast.


* subversion/libsvn_delta/text_delta.c

  (apply_window): Add typecasts.


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@841834 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/libsvn_fs/strings-table.c
@@ -218,7 +218,7 @@ string_read (svn_fs_t *fs,
   svn_fs__clear_dbt (&result);
   result.data = buf;
   result.ulen = *len;
-  result.doff = offset;
+  result.doff = (u_int32_t)offset;
   result.dlen = *len;
   result.flags |= (DB_DBT_USERMEM | DB_DBT_PARTIAL);
   db_err = cursor->c_get (cursor, &query, &result, DB_CURRENT);
